Fix breadcrumbs helper.
Adds missing overrides to ensure bootstrap classes are added for all
crumb insertion methods.

Ensures that the `aria-current` attribute is only applied to one crumb,
either the last one (default), or the last one with a link, and that for
crumbs with links the attribute is set on the link.

refs #223